backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "vendor-recovery-room-api"
version = "0.1.0"
description = "Vendor Recovery Room API"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.32.0",
  "sqlalchemy[asyncio]>=2.0.36",
  "asyncpg>=0.30.0",
  "alembic>=1.14.0",
  "pydantic>=2.10.0",
  "pydantic-settings>=2.6.0",
  "python-multipart>=0.0.17",
  "argon2-cffi>=23.1.0",
  "python-jose[cryptography]>=3.3.0",
  "email-validator>=2.2.0",
  "slowapi>=0.1.9",
  "httpx>=0.28.0",
  "boto3>=1.35.0",
  "stripe>=11.0.0",
  "python-dateutil>=2.9.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
  "pytest-asyncio>=0.24.0",
  "ruff>=0.8.0",
  "aiosqlite>=0.20.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]

[tool.ruff]
target-version = "py312"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
ignore = ["B008"]
